Component logic for ColumnContextMenu added to a React hook useColumnContextMenu
